add fcma test, modify run-tests.sh for testing out MPI processes and collecting the right coverage
