--- examples/fcma/requirements.txt
@@ -1,2 +1,0 @@
-mpi4py
-sklearn
